requirements.txt
black==23.3.0
click==8.1.3
Flask==2.2.3
flask_cors==3.0.10
flask_sqlalchemy==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pip==23.0.1
pg8000==1.29.4
python-dotenv==1.0.0
setuptools==67.6.1
sqlalchemy==2.0.9
Werkzeug==2.2.3
wheel==0.40.0
